Extract PageHeader.Title and PageHeader.Image into standalone components
