+++ apps/ff-site/src/app/_components/PageHeader.tsx
@@ -1,7 +1,7 @@
 import {
   PageHeader as SharedPageHeader,
   type PageHeaderProps as SharedPageHeaderProps,
-} from '@filecoin-foundation/ui/PageHeader'
+} from '@filecoin-foundation/ui/PageHeader/PageHeader'
 
 import { SectionDivider } from '@/components/SectionDivider'
