+++ packages/ui/src/PageHeader/index.ts
@@ -1,0 +1,5 @@
+import { PageHeader } from './PageHeader'
+import { PageHeaderImage } from './PageHeaderImage'
+import { PageHeaderTitle } from './PageHeaderTitle'
+
+export { PageHeader, PageHeaderImage, PageHeaderTitle }
